Run tests from console (using node-qunit), convert to tests to CommonJS
