+++ package.json
@@ -26,7 +26,7 @@
   "devDependencies": {
     "jshint": "^2.5.2",
     "nodeunit": "^0.9.0",
-    "qunit": "~0.6",
+    "qunit": "^0.7.1",
     "underscore": "~1.6.0"
   }
 }
